move logic into service method
