+++ backend/src/server/handlers/local.rs
@@ -219,41 +219,8 @@ pub async fn detect_markers(
     Query(DetectMarkersQuery { threshold }): Query<DetectMarkersQuery>,
     State(state): State<Arc<AppState>>,
 ) -> Result<Json<Vec<MarkerDto>>, AppError> {
-    let video = state.database.get_video(&id).await?;
-    if let Some(video) = video {
-        let threshold = threshold.unwrap_or(0.4);
-        let timestamps = scene_detection::detect_scenes(
-            &video.file_path,
-            threshold,
-            state.ffmpeg_location.clone(),
-        )
-        .await?;
-        let markers = scene_detection::detect_markers(timestamps, video.duration);
-
-        let mut created_markers = vec![];
-        for (index, marker) in markers.into_iter().enumerate() {
-            let preview_generator: PreviewGenerator = state.clone().into();
-            let preview_image = preview_generator
-                .generate_preview(&video.id, &video.file_path, marker.start)
-                .await?;
-            let db_marker = state
-                .database
-                .create_new_marker(CreateMarker {
-                    video_id: video.id.clone(),
-                    title: "Untitled".to_string(),
-                    start: marker.start,
-                    end: marker.end,
-                    preview_image_path: Some(preview_image.to_string()),
-                    index_within_video: index as i64,
-                    video_interactive: video.interactive,
-                })
-                .await?;
-            info!("created marker {db_marker:?}");
-            created_markers.push(db_marker.into());
-        }
-
-        Ok(Json(created_markers))
-    } else {
-        Err(AppError::StatusCode(StatusCode::NOT_FOUND))
-    }
+    let created_markers =
+        scene_detection::find_and_persist_markers(&id, threshold.unwrap_or(0.4), state.clone())
+            .await?;
+    Ok(Json(created_markers))
 }
