+++ backend/src/service/scene_detection.rs
@@ -1,8 +1,14 @@
+use std::sync::Arc;
+
+use clip_mash_types::{CreateMarker, MarkerDto};
+use color_eyre::eyre::bail;
 use lazy_static::lazy_static;
 use regex::Regex;
 use tracing::{debug, info};
 
 use super::commands::ffmpeg::{Ffmpeg, FfmpegLocation};
+use crate::server::handlers::AppState;
+use crate::service::preview_image::PreviewGenerator;
 use crate::Result;
 
 lazy_static! {
@@ -11,16 +17,16 @@ lazy_static! {
 
 const MIN_MARKER_DURATION: f64 = 5.0;
 
-pub async fn detect_scenes(
+async fn detect_scenes(
     input: &str,
     threshold: f64,
-    ffmpeg_location: FfmpegLocation,
+    ffmpeg_location: &FfmpegLocation,
 ) -> Result<Vec<f64>> {
     info!(
         "detecting scenes in input {} with threshold {}",
         input, threshold
     );
-    let output = Ffmpeg::new(&ffmpeg_location, "-")
+    let output = Ffmpeg::new(ffmpeg_location, "-")
         .input(input)
         .format("null")
         .video_filter(format!("select='gt(scene,{threshold})',showinfo'"))
@@ -43,12 +49,12 @@ pub async fn detect_scenes(
 }
 
 #[derive(Debug, Clone)]
-pub struct DetectedMarker {
+struct DetectedMarker {
     pub start: f64,
     pub end: f64,
 }
 
-pub fn detect_markers(mut timestamps: Vec<f64>, total_duration: f64) -> Vec<DetectedMarker> {
+fn detect_markers(mut timestamps: Vec<f64>, total_duration: f64) -> Vec<DetectedMarker> {
     if timestamps.is_empty() {
         return vec![DetectedMarker {
             start: 0.0,
@@ -78,4 +84,42 @@ pub fn detect_markers(mut timestamps: Vec<f64>, total_duration: f64) -> Vec<Dete
         }
     }
     markers
+}
+
+pub async fn find_and_persist_markers(
+    video_id: &str,
+    threshold: f64,
+    state: Arc<AppState>,
+) -> Result<Vec<MarkerDto>> {
+    // let threshold = threshold.unwrap_or(0.4);
+    let video = state.database.get_video(video_id).await?;
+    if video.is_none() {
+        bail!("no video found for id {}", video_id);
+    }
+    let video = video.unwrap();
+    let timestamps = detect_scenes(&video.file_path, threshold, &state.ffmpeg_location).await?;
+    let markers = detect_markers(timestamps, video.duration);
+
+    let mut created_markers = vec![];
+    for (index, marker) in markers.into_iter().enumerate() {
+        let preview_generator: PreviewGenerator = state.clone().into();
+        let preview_image = preview_generator
+            .generate_preview(&video.id, &video.file_path, marker.start)
+            .await?;
+        let db_marker = state
+            .database
+            .create_new_marker(CreateMarker {
+                video_id: video.id.clone(),
+                title: "Untitled".to_string(),
+                start: marker.start,
+                end: marker.end,
+                preview_image_path: Some(preview_image.to_string()),
+                index_within_video: index as i64,
+                video_interactive: video.interactive,
+            })
+            .await?;
+        info!("created marker {db_marker:?}");
+        created_markers.push(db_marker.into());
+    }
+    Ok(created_markers)
 }
